Fix IZPACK-703: NPE check for guiPrefsElement
According to XSD, <guiprefs> is optional but Packager unconditionally
tries to get it's child, <splash> element, when writing manifest file
and fails with NPE if <guiprefs> is missing.

There is a workaround - provide a dummy guiprefs. But when writing a
headless installer, there is no sense in <guiprefs> at all.

So makining it a mandatory element in XSD is not the best choice. Just
add a check for a missing <guiprefs>.