IZPACK-815: Sub menus are not created in Gnome3
Related: http://lists.fedoraproject.org/pipermail/devel/2011-August/155342.html

Gnome3 does not follow the freedesktop specifications anymore! There is a
mismatch between the path that gnome3 looks for the menu and the path defined in
the freedesktop specification. That's why it worked fine in Gnome2 (and KDE 4.x)
but not in Gnome3.

Solution is to write the .menu file in ~/.config/menus/application-gnome-merged/

Thanks to Jyoti Tripathi for the help!